// src/types/FTTHPreorder.ts
export interface FTTHPreorder {
    ID: number;
    Eshop_ID: number;
    Postal_Code: string;
    Province: string;
    City: string;
    Tracking_Code: string;
    Lat: number;
    Long: number;
    Created_Date: string;
    FTTH_ID: number | null;
    Product_Name: string;
    FAT_ID: number | null;
  }
  